requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.7.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
